requirements.txt
APScheduler==3.6.3
async-lru==1.0.2
certifi==2020.12.5
charset-normalizer==2.0.7
flake8==4.0.1
hickory==1.0.0
idna==3.3
isort==5.10.1
mccabe==0.6.1
pyaes==1.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
Pyrogram==1.2.8
PySocks==1.7.1
pyTelegramBotAPI==4.2.0
python-dotenv==0.17.0
python-telegram-bot==13.4.1
pytz==2021.1
requests==2.26.0
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.7
